Customer Churn Analysis using ML - Blog
Customer retention is critical to a company's success. Customers are a company's most precious asset, impacting business today and growing in value over time as they continue to engage in products and services. Customer churn can be costly for both new and established businesses. The true cost of churn is much larger than what corporate leaders assume. Not only does this result in lost revenue in the short term, but it also means that your team will have to double down on gaining new customers to replenish those revenue streams in order to assure future success. It is widely acknowledged that acquiring a new customer might cost up to 5 times as much as keeping an existing one.